Move config and toml format parsers to their own formats modules
